combine files into lib.rs
