+++ Cargo.toml
@@ -2,7 +2,10 @@
 name = "ftapi"
 version = "0.1.0"
 edition = "2021"
-authors= ["bagmeg"]
+authors = ["bagmeg"]
+license = "Apache-2.0"
+readme = "README.md"
+repository = "https://github.com/bagmeg/ftapi"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
